backend/requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
openai
google-generativeai
pydantic
python-multipart
pymupdf
python-docx
